backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.13.0
python-multipart>=0.0.9
pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
cryptography>=42.0.0
aiofiles>=23.0.0
